+++ GeneralsMD/Code/GameEngine/Include/GameLogic/Module/QueueProductionExitUpdate.h
@@ -58,6 +58,13 @@ public:
 		m_initialBurst = 0;
 	}
 
+	virtual Bool getProductionExitPointsInModelSpace( Coord3D& unitCreatePoint, Coord3D& naturalRallyPoint ) const
+	{
+		unitCreatePoint = m_unitCreatePoint;
+		naturalRallyPoint = m_naturalRallyPoint;
+		return TRUE;
+	}
+
 	static void buildFieldParse(MultiIniFieldParse& p) 
 	{
     UpdateModuleData::buildFieldParse(p);