+++ GeneralsMD/Code/GameEngine/Include/GameLogic/Module/SupplyCenterProductionExitUpdate.h
@@ -51,6 +51,13 @@ public:
 		m_naturalRallyPoint.zero();
 	}
 
+	virtual Bool getProductionExitPointsInModelSpace( Coord3D& unitCreatePoint, Coord3D& naturalRallyPoint ) const
+	{
+		unitCreatePoint = m_unitCreatePoint;
+		naturalRallyPoint = m_naturalRallyPoint;
+		return TRUE;
+	}
+
 	static void buildFieldParse(MultiIniFieldParse& p) 
 	{
     UpdateModuleData::buildFieldParse(p);